requirements.txt
torch
open_clip_torch
faiss-cpu
Pillow
numpy
streamlit
